Refactor lsi
